enh/theme: Move Todo interface from types to todos/state
